Crate dokany

Re-exports

pub use dokany_sys as sys;

Structs

AccessMask

A value that defines standard, specific, and generic rights.

FileSystemFlags

Capability flags for a filesystem.

FillFindData

Supply a FindData entry for directory listing.

FindData

A dir entry

MainResult

The error/result for mounting a filesystem.

OptionFlags

Options to configure the driver mount

Options

Options for a filesystem

WriteWideCStringCell

A cell that wraps an unitialzed wide c string buffer, tracking its initialization.

Traits

AsWide

Implemented for types that can be converted into wide char arrays.

FileSystem

The trait a type must implement to serve as a file system

Functions

driver_version

Get the driver version.

init

Initialize the library, if needed.

main

Mount and run a filesystem from the given options an mount object.

remove_mount_point

Unmount a drive from the given path.

shutdown⚠

Shutdown the api and release all resources.

unmount

Unmount the drive from the given drive letter.

version

Get the version.