Trait dokany::AsWide

pub trait AsWide {
    type Iter: Iterator<Item = u16>;

    // Required method
    fn as_wide(&self) -> Self::Iter;
}

Implemented for types that can be converted into wide char arrays.

Required Associated Types

type Iter: Iterator<Item = u16>

Required Methods

fn as_wide(&self) -> Self::Iter

Get an iterator over wide chars.

Depending on where this is passed, it is a logic error to have a NUL wide char.